requirements.txt
bs4==0.0.2
click==6.7
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
flask-login===0.6.3
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.4.0
SQLAlchemy==1.3.6
Werkzeug==1.0.1
markdown==3.4.4
pyyaml==6.0.1
python-avatars==1.3.1